+++ www/js/util.js
@@ -700,6 +700,7 @@ function register_business() {
     var confirm_password = 'pass';
     var city_id = $('#business_register-city_select').val().trim();
     var address = $('#business_register-address').val().trim();
+    var profile_image = profile_image_link;
     var business_category = '';
     var new_category = '';
 
@@ -765,6 +766,11 @@ function register_business() {
         return false;
     }
 
+    if (!profile_image) {
+        myApp.alert('Please provide profile picture.');
+        return false;
+    }
+
     // business_category = business_category.slice(0, -1);
 
     myApp.showIndicator();
@@ -787,6 +793,7 @@ function register_business() {
             user_type: 'Business',
             phone: phone,
             new_category: new_category,
+            profile_image: profile_image,
         },
     }).done(function(res) {
         myApp.hideIndicator();
@@ -1631,9 +1638,9 @@ function loadUsersPageContent(user_id) {
                                     '</a>'+
                                     '<div class="card-footer no-border like_share pad0" style="width: 40%;">'+
                                     '<a href="javascript:void(0);" data-liked="0" class=""><i onclick="feedShareStatusChng('+value.feed_id+')" data-title="'+title+'" data-image_link="'+share_image_link+'" class="material-icons white_heart share_feeds_'+value.feed_id+'">share</i></a>';
-                                    // if (value.user_id == token.id) {
+                                    if (value.user_id == token.id) {
                                         save_feeds_html += '<a href="javascript:void(0);" data-liked="0" onclick="delete_saved('+value.rel_id+')" class=""><i class="material-icons white_heart">delete</i></a>';
-                                    // }
+                                    }
                     save_feeds_html += '</div>'+
                                     '</div>';
                 }
@@ -1646,9 +1653,9 @@ function loadUsersPageContent(user_id) {
                                     '</a>'+
                                     '<div class="card-footer no-border like_share pad0" style="width: 40%;">'+
                                     '<a href="javascript:void(0);" data-liked="0" class=""><i onclick="feedShareStatusChng('+value.feed_id+')" data-title="'+title+'" data-image_link="'+share_image_link+'" class="material-icons white_heart share_feeds_'+value.feed_id+'">share</i></a>';
-                                    // if (value.user_id == token.id) {
+                                    if (value.user_id == token.id) {
                                         save_feeds_html += '<a href="javascript:void(0);" data-liked="0" onclick="delete_saved('+value.rel_id+')" class=""><i class="material-icons white_heart">delete</i></a>';
-                                    // }
+                                    }
                     save_feeds_html += '</div>'+
                                     '</div>';
                 }
